+++ team/m3/backend/server.js
@@ -2,13 +2,9 @@ import express from 'express';
 import path from 'path';
 import { fileURLToPath } from 'url';
 import { dirname } from 'path';
-
-
-import TokenRoutes from './routers/tokenRoutes.js'
 import tileCoordRouter from "../js/backend/routers/TileCoordRouter"
 import tileRouter from "../js/backend/routers/TileRouter"
 import GridRouter from './routers/GridRouter';
-
 
 class Server {
     constructor(){
@@ -27,9 +23,6 @@ class Server {
 
     //Each person adds their routes here
     setupRoutes() {
-
-        this.app.use("/tokens", TokenRoutes)
-
         this.app.use('/tiles', tileRouter);
         this.app.use('/tileCoordinates', tileCoordRouter);
         this.app.use('/grid', GridRouter);
